Ready for running experiments on other PC
basic experiments are now performing the same operations on all DS

All data structures also are capable of returning the amount of non-static
memory they take up.

made it convenient to run both persistent tests and normal tests with a single
bash file. These tests have their own files which are generated in the results
folder.

+++ src/SLL.h
@@ -29,7 +29,7 @@ Generic SLLGetElement(SLL *sll, int64_t index);
 uint64_t SLLGetSize(SLL *sll);
 
 // actual memory size
-uint64_t SLLGetMemSize(SLL *sll);
+uintptr_t SLLGetMemSize(SLL *sll);
 
 void SLLPrint(FILE *out, SLL *sll);
